Refactor onboarding celebration reveal
- Extract CelebrationReveal and CelebrationParticles from Onboarding.jsx
- Preserve celebration timing, reduced-motion behavior, and prop contract
- Leave handleFinish, auth metadata ordering, routing, Supabase writes, and validation untouched
